going back to LD_PRELOAD approach
The issue is that it is more or less impossible to
intercept a call to _init in a shared library.
LD_PRELOAD and hooking __libc_start_main gives us a
chance to enable SECCOMP before untrusted constructor
functions run.

Note that we still need to be worried about executables
that have a nonstandard entry point, and/or that don't
call __libc_start_main.  Assuming that our untrusted
code has been compiled from source, that should not
be an issue.