yet another reimplementation using klibc
I'm convinced that this approach will never work using
glibc: it's just too big and complicated and uses codepaths
that invoke all kinds of system calls.

New approach is to use klibc, and to wrap the untrusted
main function by renaming it (using objcopy) and then
linking in our own main() which enables seccomp.
The current sticking point is trying to convince
the memory allocator to leave a large arena so that memory
can be allocated without having to call mmap or sbrk.